[modify] refactor w/ RuboCop

+++ app/controllers/concerns/gws/crud_filter.rb
@@ -44,11 +44,9 @@ module Gws::CrudFilter
 
   def edit
     raise "403" unless @item.allowed?(:edit, @cur_user, site: @cur_site)
-    if @item.is_a?(Cms::Addon::EditLock)
-      unless @item.acquire_lock
-        redirect_to action: :lock
-        return
-      end
+    if @item.is_a?(Cms::Addon::EditLock) && !@item.acquire_lock
+      redirect_to action: :lock
+      return
     end
     render
   end
@@ -130,11 +128,9 @@ module Gws::CrudFilter
     entries.each do |item|
       if item.allowed?(:delete, @cur_user, site: @cur_site)
         item.attributes = fix_params
-        if item.is_a?(Gws::User)
-          if item.deletion_unlocked? && item.disabled?
-            item.destroy
-            next
-          end
+        if item.is_a?(Gws::User) && item.deletion_unlocked? && item.disabled?
+          item.destroy
+          next
         end
         next if item.disable
       else
